+++ src/Junction.php
@@ -4,6 +4,9 @@ namespace Weap\Junction;
 
 use Illuminate\Support\Facades\Route;
 
+/**
+ * @deprecated Junction::resource are replaced by Route::junctionResource.
+ */
 class Junction
 {
     /**
@@ -11,20 +14,8 @@ class Junction
      * @param $controller
      * @return void
      */
-    public static function resource($uri, $controller, $only = ['index', 'show', 'store', 'update', 'destroy', 'action']): void
+    public static function resource($uri, $controller, $only = ['index', 'indexPost', 'store', 'show', 'showPost', 'update', 'destroy', 'action']): void
     {
-        Route::apiResource($uri, $controller)->only(array_diff($only, ['action']));
-
-        if (in_array('index', $only)) {
-            Route::post($uri . '/index', $controller. '@index');
-        }
-
-        if (in_array('show', $only)) {
-            Route::post($uri . '/{id}/show', $controller. '@show');
-        }
-
-        if (in_array('action', $only)) {
-            Route::put($uri, $controller . '@action');
-        }
+        Route::junctionResource($uri, $controller)->only($only);
     }
 }